Add new addAppliant form page
