--- frontend/src/pages/Dashboard/ApplicantTracking/sections/AddApplicant.jsx
@@ -1,79 +1,0 @@
-// import { ArrowLeftIcon } from "@radix-ui/react-icons";
-// import { Input } from "/src/components/ui/input";
-// import { Button } from "/src/components/ui/button";
-// import { Label } from "/src/components/ui/label";
-
-// const AddApplicant = () => {
-//   return (
-//     <section className="w-full bg-lilac-pale rounded-t-[2.5rem] p-8 flex flex-col">
-//       <div className="my-4">
-//         <span className="flex items-center justify-start cursor-pointer w-fit space-x-4">
-//           <ArrowLeftIcon className="h-8 w-8" color="gray" />
-//           <h1 className="inline text-4xl font-bold text-lilac font-lato   ">
-//             Add New Applicant
-//           </h1>
-//         </span>
-//       </div>
-//       <div className="bg-white flex-grow rounded-xl shadow-lg p-8">
-//         <form className="flex h-full w-full gap-8">
-//           <div className="w-1/2">
-//             <div className="flex gap-4 mb-2 ">
-//               <span className="w-1/2">
-//                 <Label>First Name</Label>
-//                 <Input className="h-10 md:h-12 bg-zinc-100 focus-visible:ring-lilac" />
-//               </span>
-//               <span className="w-1/2">
-//                 <Label>Middle Name</Label>
-//                 <Input className="h-10 md:h-12 bg-zinc-100 focus-visible:ring-lilac" />
-//               </span>
-//             </div>
-//             <div className="flex gap-4 mb-2">
-//               <span className="w-1/2">
-//                 <Label>Last Name</Label>
-//                 <Input className="h-10 md:h-12 bg-zinc-100 focus-visible:ring-lilac" />
-//               </span>
-//               <span className="w-1/2">
-//                 <Label>Suffix</Label>
-//                 <Input className="h-10 md:h-12 bg-zinc-100 focus-visible:ring-lilac" />
-//               </span>
-//             </div>
-//             <div className="flex gap-4 mb-2">
-//               <span className="flex w-1/2 gap-4 ">
-//                 <span className="w-1/4">
-//                   <Label>Age</Label>
-//                   <Input className="h-10 md:h-12 bg-zinc-100   focus-visible:ring-lilac" />
-//                 </span>
-//                 <span className="w-3/4">
-//                   <Label>BirthDate</Label>
-//                   <Input className="h-10 md:h-12 bg-zinc-100   focus-visible:ring-lilac" />
-//                 </span>
-//               </span>
-//               <span className="w-1/2">
-//                 <Label>Sex</Label>
-//                 <Input className="h-10 md:h-12 bg-zinc-100 focus-visible:ring-lilac" />
-//               </span>
-//             </div>
-//           </div>
-//           <div className="w-1/2">
-//             <div className="mb-2">
-//               <Label>Email Address</Label>
-//               <Input className="h-10 md:h-12 bg-zinc-100   focus-visible:ring-lilac" />
-//             </div>
-//             <div className="flex gap-4 mb-2 ">
-//               <span className="w-1/2">
-//                 <Label>Mobile Number</Label>
-//                 <Input className="h-10 md:h-12 bg-zinc-100 focus-visible:ring-lilac" />
-//               </span>
-//               <span className="w-1/2">
-//                 <Label>Landline Number</Label>
-//                 <Input className="h-10 md:h-12 bg-zinc-100 focus-visible:ring-lilac" />
-//               </span>
-//             </div>
-//           </div>
-//         </form>
-//       </div>
-//     </section>
-//   );
-// };
-
-// export default AddApplicant;
