--- frontend/src/pages/Dashboard/ApplicantTracking/sections/index.jsx
@@ -1,19 +1,0 @@
-// import { useState } from "react";
-// import ViewApplicant from "./sections/ViewApplicant";
-// import AddApplicant from "./sections/AddApplicant";
-
-// const ApplicantTracking = () => {
-//   const [addApplicant, setAddApplicant] = useState(false);
-
-//   return (
-//     <div className="flex h-full">
-//       {addApplicant ? (
-//         <AddApplicant onExit={() => setAddApplicant(false)} />
-//       ) : (
-//         <ViewApplicant onAddApplicant={() => setAddApplicant(true)} />
-//       )}
-//     </div>
-//   );
-// };
-
-// export default ApplicantTracking;
